Assignment #4
testdominion.c is written, it is a complete random tester for dominion.c
    By complete of course I mean it plays full games, it does not 100%
    test dominion.

There are makefile rules for this as well which output test results to
testdominion.out

Included are also the results from running two other student's dominion.c
code.

There is an analysis of the differences (or lack of) in differential.txt.